requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-crispy-forms==1.12.0
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
tk==0.1.0
toml==0.10.2